+++ sdk/storage/Azure.Storage.Blobs/src/BlobServiceClient.cs
@@ -136,7 +136,7 @@ namespace Azure.Storage.Blobs
         /// Determines whether the client is able to generate a SAS.
         /// If the client is authenticated with a <see cref="StorageSharedKeyCredential"/>.
         /// </summary>
-        public bool CanGenerateSasUri => _storageSharedKeyCredential != null;
+        public bool CanGenerateAccountSasUri => _storageSharedKeyCredential != null;
 
         #region ctors
         /// <summary>
@@ -193,7 +193,7 @@ namespace Azure.Storage.Blobs
             _customerProvidedKey = options.CustomerProvidedKey;
             _clientSideEncryption = options._clientSideEncryptionOptions?.Clone();
             _encryptionScope = options.EncryptionScope;
-            _storageSharedKeyCredential = StorageSharedKeyCredential.ParseConnectionString(connectionString);
+            _storageSharedKeyCredential = conn.Credentials as StorageSharedKeyCredential;
             BlobErrors.VerifyHttpsCustomerProvidedKey(_uri, _customerProvidedKey);
             BlobErrors.VerifyCpkAndEncryptionScopeNotBothSet(_customerProvidedKey, _encryptionScope);
         }
@@ -1783,7 +1783,7 @@ namespace Azure.Storage.Blobs
         /// shared key credential of the client.
         ///
         /// To check if the client is able to sign a Service Sas see
-        /// <see cref="CanGenerateSasUri"/>.
+        /// <see cref="CanGenerateAccountSasUri"/>.
         ///
         /// For more information, see
         /// <see href="https://docs.microsoft.com/en-us/rest/api/storageservices/create-account-sas">
@@ -1824,7 +1824,7 @@ namespace Azure.Storage.Blobs
         /// shared key credential of the client.
         ///
         /// To check if the client is able to sign a Service Sas see
-        /// <see cref="CanGenerateSasUri"/>.
+        /// <see cref="CanGenerateAccountSasUri"/>.
         ///
         /// For more information, see
         /// <see href="https://docs.microsoft.com/en-us/rest/api/storageservices/create-account-sas">