+++ sdk/storage/Azure.Storage.Common/api/Azure.Storage.Common.netstandard2.0.cs
@@ -17,8 +17,6 @@ namespace Azure.Storage
         public StorageSharedKeyCredential(string accountName, string accountKey) { }
         public string AccountName { get { throw null; } }
         protected static string ComputeSasSignature(Azure.Storage.StorageSharedKeyCredential credential, string message) { throw null; }
-        [System.ComponentModel.EditorBrowsableAttribute(System.ComponentModel.EditorBrowsableState.Never)]
-        public static Azure.Storage.StorageSharedKeyCredential ParseConnectionString(string connectionString) { throw null; }
         public void SetAccountKey(string accountKey) { }
     }
     [System.Runtime.InteropServices.StructLayoutAttribute(System.Runtime.InteropServices.LayoutKind.Sequential)]