+++ sdk/storage/Azure.Storage.Common/src/Shared/Constants.cs
@@ -539,6 +539,12 @@ namespace Azure.Storage
                 Sas.Permissions.Move,
                 Sas.Permissions.Execute
             };
+
+            /// <summary>
+            /// List of ports used for path style addressing.
+            /// Copied from Microsoft.Azure.Storage.Core.Util
+            /// </summary>
+            internal static readonly int[] PathStylePorts = { 10000, 10001, 10002, 10003, 10004, 10100, 10101, 10102, 10103, 10104, 11000, 11001, 11002, 11003, 11004, 11100, 11101, 11102, 11103, 11104 };
         }
 
         internal static class ClientSideEncryption