+++ sdk/storage/Azure.Storage.Files.DataLake/src/DataLakeServiceClient.cs
@@ -105,7 +105,7 @@ namespace Azure.Storage.Files.DataLake
         /// Determines whether the client is able to generate a SAS.
         /// If the client is authenticated with a <see cref="StorageSharedKeyCredential"/>.
         /// </summary>
-        public bool CanGenerateSasUri => _storageSharedKeyCredential != null;
+        public bool CanGenerateAccountSasUri => _storageSharedKeyCredential != null;
 
         #region ctors
         /// <summary>
@@ -770,7 +770,7 @@ namespace Azure.Storage.Files.DataLake
         /// shared key credential of the client.
         ///
         /// To check if the client is able to sign a Service Sas see
-        /// <see cref="CanGenerateSasUri"/>.
+        /// <see cref="CanGenerateAccountSasUri"/>.
         ///
         /// For more information, see
         /// <see href="https://docs.microsoft.com/en-us/rest/api/storageservices/create-account-sas">
@@ -811,7 +811,7 @@ namespace Azure.Storage.Files.DataLake
         /// The SAS is signed by the shared key credential of the client.
         ///
         /// To check if the client is able to sign a Service Sas see
-        /// <see cref="CanGenerateSasUri"/>.
+        /// <see cref="CanGenerateAccountSasUri"/>.
         ///
         /// For more information, see
         /// <see href="https://docs.microsoft.com/en-us/rest/api/storageservices/create-account-sas">