+++ sdk/storage/Azure.Storage.Blobs/CHANGELOG.md
@@ -7,8 +7,8 @@
 - Fixed bug where Blobs SDK coudn't handle SASs with start and expiry time in format other than yyyy-MM-ddTHH:mm:ssZ.
 - Added ability to set Position on streams created with BlobBaseClient.OpenRead().
 - Added CanGenerateSasUri property, GenerateSasUri() to BlobBaseClient, BlobClient, BlockBlobClient, AppendBlobClient, PageBlobClient and BlobContainerClient.
-- Added CanGenerateSasUri property, GenerateSasUri() to BlobServiceClient.
-- Deprecated proeprty BlobSasBuilder.Version and when generating SAS will always use the latest storage SAS version of the Storage service version supported.
+- Added CanAccountGenerateSasUri property, GenerateAccountSasUri() to BlobServiceClient.
+- Deprecated property BlobSasBuilder.Version, so when generating SAS will always use the latest Storage Service SAS version.
 
 ## 12.7.0-preview.1 (2020-09-30)
 - Added support for service version 2020-02-10.